create io module, move config, visualize there. abstract `Frame` and `Instance` classes into separate modules within `io`
