+++ biogtr/inference/track.py
@@ -1,8 +1,8 @@
 """Script to run inference and get out tracks."""
 
-from biogtr.config import Config
+from biogtr.io.config import Config
 from biogtr.models.gtr_runner import GTRRunner
-from biogtr.data_structures import Frame
+from biogtr.io.frame import Frame
 from omegaconf import DictConfig
 from pathlib import Path
 from pprint import pprint
