+++ biogtr/io/__init__.py
@@ -1,0 +1,6 @@
+"""Module containing input/output data structures for easy storage and manipulation."""
+
+from biogtr.io.frame import Frame
+from biogtr.io.instance import Instance
+
+# TODO: expose config without circular import error from biogtr.io.config import Config
