+++ biogtr/training/losses.py
@@ -1,7 +1,7 @@
 """Module containing different loss functions to be optimized."""
 
-from biogtr.data_structures import Frame
-from biogtr.models.model_utils import get_boxes, get_times
+from biogtr.io.frame import Frame
+from biogtr.models.model_utils import get_boxes_times
 from torch import nn
 from typing import List, Tuple
 import torch
