+++ tests/test_data_structures.py
@@ -1,6 +1,7 @@
 """Tests for Instance, Frame, and TrackQueue Object"""
 
-from biogtr.data_structures import Instance, Frame
+from biogtr.io.frame import Frame
+from biogtr.io.instance import Instance
 from biogtr.inference.track_queue import TrackQueue
 import torch
